+++ monopyly/static/js/autocomplete_credit_transaction.js
@@ -21,7 +21,12 @@ import { AutocompleteBox } from './modules/autocomplete_input.js';
     // Use a delegated event handler to use autocomplete
     $('form').on('focus', '.autocomplete input', function() {
       const inputElement = this;
-      autocompleteBox = new AutocompleteBox(inputElement);
+      const inputName = $(inputElement).attr("name");
+      if (inputName.endsWith("tags")) {
+        autocompleteBox = new TagAutocompleteBox(inputElement);
+      } else {
+        autocompleteBox = new AutocompleteBox(inputElement);
+      }
       // Identify the merchant input box (used when suggesting notes)
       const $merchant = $('form input#merchant');
       autocompleteAjaxRequest(autocompleteBox, inputElement, $merchant);
@@ -45,3 +50,46 @@ import { AutocompleteBox } from './modules/autocomplete_input.js';
 
 })();
 
+
+/**
+ * A class for performing autocomplete on tags inputs.
+ */
+class TagAutocompleteBox extends AutocompleteBox {
+
+  /**
+   * Create the autcomplete box for tags.
+   */
+  constructor(inputElement) {
+    super(inputElement);
+  }
+
+  /**
+   * Populate the last tag in the input element using the chosen suggestion.
+   */
+  fill(suggestion) {
+    const lastTagName = suggestion.textContent;
+    const tagArray = this.#separateTags(this.inputElement.value);
+    // Pad the added tag with a space (if it is not the first/only element)
+    let padding;
+    if (tagArray.length == 1) {
+      padding = "";
+    } else {
+      padding = " ";
+    }
+    tagArray[tagArray.length-1] = padding + lastTagName;
+    this.inputElement.value = tagArray.join(",");
+  }
+
+  /**
+   * Refesh autocomplete suggestions using only the last input tag.
+   */
+  refreshMatches(userInput) {
+    const autocompleteSegment = this.#separateTags(userInput).pop().trim();
+    super.refreshMatches(autocompleteSegment);
+  }
+
+  #separateTags(userInput) {
+    return userInput.split(",");
+  }
+
+}