+++ CHANGELOG.md
@@ -95,5 +95,6 @@
 
 ### 1.2.4 (in progress)
 
+- Add an enhanced autocompletion for assigning tags to credit transactions
 - Leverage SQLAlchemy 2.0 `DeclarativeBase` class and `declared_attr` decorators
 - Simplify `Model` base to avoid explicitly defining column attributes