+++ CHANGELOG.md
@@ -93,9 +93,11 @@
 - Use metaclasses for database handler to avoid stacking classmethod and property (deprecated in Python 3.11)
 
 
-### 1.2.4 (in progress)
+## 1.3.0 (in progress)
 
-- Add an enhanced autocompletion for assigning tags to credit transactions
+- Add a merchant field and tags to bank transactions
+- Add an enhanced autocompletion for assigning transaction tags
 - Swap the README instructions for the 'About' page (and move the story to a separate page)
 - Leverage SQLAlchemy 2.0 `DeclarativeBase` class and `declared_attr` decorators
 - Simplify `Model` base to avoid explicitly defining column attributes
+- Update the database backup script to work as part of the package